pyproject.toml
[project]
name = "data-contract-enforcer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.86.0",
    "gitpython>=3.1.46",
    "jsonschema>=4.26.0",
    "langsmith>=0.7.23",
    "numpy>=2.3.5",
    "openai>=2.30.0",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "soda-core>=4.3.0",
    "ydata-profiling>=4.18.1",
]
